+++ src/projectGeneration/commands.ts
@@ -2,69 +2,74 @@ import * as vscode from "vscode";
 import { scaffoldProjectRequest } from ".";
 import { registerCommandWithLogging } from "../commands";
 import { ResourceLoader } from "../loaders";
+import { CCloudResourceLoader } from "../loaders/ccloudResourceLoader";
 import { CCloudFlinkComputePool } from "../models/flinkComputePool";
 import { KafkaCluster } from "../models/kafkaCluster";
+import { CCloudOrganization } from "../models/organization";
 import { KafkaTopic } from "../models/topic";
 import { showErrorNotificationWithButtons } from "../notifications";
 import { removeProtocolPrefix } from "../utils/bootstrapServers";
 
-export function registerProjectGenerationCommands(context: vscode.ExtensionContext) {
-  context.subscriptions.push(
+export function registerProjectGenerationCommands(
+  context: vscode.ExtensionContext,
+): vscode.Disposable[] {
+  return [
     registerCommandWithLogging("ccloud.scaffoldProject", async () => {
       await scaffoldProjectRequest();
     }),
-  );
-
-  context.subscriptions.push(
     registerCommandWithLogging("ccloud.scaffoldResource", async (resource: ResourceLoader) => {
       await resourceScaffoldProjectRequest(resource);
     }),
-  );
+  ];
 }
 
 export async function resourceScaffoldProjectRequest(resource: ResourceLoader) {
   if (resource instanceof KafkaCluster) {
-    const clusterId = resource.id;
-    const organizationId = (resource as any).parent?.id;
-    const clusterName = resource.name;
-    const bootstrap = (resource as any).bootstrap;
-    const bootstrapWithoutProtocol = removeProtocolPrefix(bootstrap);
-
-    await scaffoldProjectRequest({
-      templateCollection: "vscode",
-      templateName: "kafka",
-      clusterId: String(clusterId),
-      organizationId: String(organizationId),
-      clusterName: String(clusterName),
-      bootstrap: String(bootstrapWithoutProtocol),
-    });
+    const bootstrapServers: string = removeProtocolPrefix(resource.bootstrapServers);
+    return await scaffoldProjectRequest(
+      {
+        bootstrapServer: bootstrapServers,
+        ccBootstrapServer: bootstrapServers,
+        templateType: "kafka",
+      },
+      "cluster",
+    );
   } else if (resource instanceof KafkaTopic) {
-    const topicName = resource.name;
-    const clusterId = (resource as any).parent?.id;
-    const organizationId = ((resource as any).parent?.parent as any)?.id;
-    const bootstrap = (resource as any).parent?.bootstrap;
-    const bootstrapWithoutProtocol = removeProtocolPrefix(bootstrap);
-
-    await scaffoldProjectRequest({
-      templateCollection: "vscode",
-      templateName: "kafka",
-      topicName: String(topicName),
-      clusterId: String(clusterId),
-      organizationId: String(organizationId),
-      bootstrap: String(bootstrapWithoutProtocol),
-    });
+    const clusters = await ResourceLoader.getInstance(
+      resource.connectionId,
+    ).getKafkaClustersForEnvironmentId(resource.environmentId);
+    const cluster = clusters.find((c) => c.id === resource.clusterId);
+    if (!cluster) {
+      showErrorNotificationWithButtons(
+        `Unable to find Kafka cluster for topic "${resource.name}".`,
+      );
+      return;
+    }
+    const bootstrapServers: string = removeProtocolPrefix(cluster.bootstrapServers);
+    return await scaffoldProjectRequest(
+      {
+        bootstrapServer: bootstrapServers,
+        ccBootstrapServer: bootstrapServers,
+        ccTopic: resource.name,
+        topic: resource.name,
+        templateType: "kafka",
+      },
+      "topic",
+    );
   } else if (resource instanceof CCloudFlinkComputePool) {
-    const computePoolId = resource.id;
-    const organizationId = (resource as any).parent?.id;
-    const computePoolName = resource.name;
-
-    await scaffoldProjectRequest({
-      templateCollection: "vscode",
-      templateName: "flink",
-      computePoolId: String(computePoolId),
-      organizationId: String(organizationId),
-      computePoolName: String(computePoolName),
-    });
+    const organization: CCloudOrganization | undefined =
+      await CCloudResourceLoader.getInstance().getOrganization();
+    return await scaffoldProjectRequest(
+      {
+        ccEnvironmentId: resource.environmentId,
+        ccOrganizationId: organization?.id,
+        cloudRegion: resource.region,
+        cloudProvider: resource.provider,
+        ccComputePoolId: resource.id,
+        templateType: "flink",
+      },
+      "compute pool",
+    );
   } else {
     await showErrorNotificationWithButtons("Scaffolding is not supported for this resource type", {
       OK: () => {},
